+++ src/test/kotlin/racingcar/WinnerTest.kt
@@ -21,21 +21,4 @@ class WinnerTest {
 
         assertThat(Winner(positions).names()).containsExactly("pobi", "honux")
     }
-
-    class Winner(private val positions: List<Positions>) {
-        fun names(): List<String> {
-            return positions.last()
-                .sortedByDescending { it.intValue() }
-                .takeIf { it.isNotEmpty() }
-                ?.run {
-                    winnerNames(this, first().intValue())
-                } ?: emptyList()
-        }
-
-        private fun winnerNames(
-            positions: List<NamedPosition>,
-            mostPosition: Int
-        ) = positions.takeWhile { it.intValue() >= mostPosition }
-            .map { it.name.stringValue() }
-    }
 }
